kaizen/requirements.txt
fastapi
nltk
openai
pydantic
python-dotenv
PyYAML
torch
uvicorn
